Remove builder code from glTF, so that all that is left is JSON stuff
The builder code has been moved into GlTFExporter
