+++ src/main/kotlin/models/formats/MaterialBuffers.kt
@@ -1,4 +1,4 @@
-package models.formats.glTF
+package models.formats
 
 import cache.utils.ColorPalette
 import java.awt.Color
